requirements.txt

aiohttp==3.7.3
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.3.0
azure-core==1.10.0
azure-data-tables==12.0.0b4
beautifulsoup4==4.9.3
boto==2.49.0
boto3==1.9.66
botocore==1.12.67
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cryptography==3.3.1
distlib==0.3.1
docutils==0.16
filelock==3.0.12
gensim==3.8.0
google-api-core==1.22.2
google-auth==1.24.0
google-cloud-core==1.5.0
google-cloud-storage==1.35.0
google-crc32c==1.1.0
google-resumable-media==1.2.0
googleapis-common-==1.52.0
idna==2.10
idna-ssl==1.1.0
importlib-metadata==3.4.0
importlib-resources==5.0.0
jmespath==0.10.0
joblib==1.0.0
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
multidict==4.7.6
nltk==3.5
numpy==1.19.2
pandas==1.1.3
pip==20.3.3
pipenv==2020.11.15
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.5
regex==2020.11.13
requests==2.25.1
rsa==4.7
s3transfer==0.1.13
scipy==1.5.2
setuptools==51.1.2.post20210110
six==1.15.0
smart-open==4.1.0
soupsieve==2.1
tqdm==4.55.1
typing-extensions==3.7.4.3
urllib3==1.24.3
virtualenv==20.3.0
virtualenv-clone==0.5.4
wheel==0.36.2
win-inet-pton==1.1.0
wincertstore==0.2
yarl==1.5.1
zipp==3.4.0
